Added new ListViewController.
